Conflicts:
	UPGRADE.asciidoc

+++ addons/packages/packetfence.spec
@@ -264,6 +264,7 @@ Requires: perl(Test::NoWarnings)
 Requires: perl(Net::UDP)
 # For managing the number of connections per device
 Requires: mod_qos
+Requires: %{real_name}-config
 
 %description -n %{real_name}
 
@@ -317,6 +318,17 @@ Summary: Replace pfcmd by a C wrapper for suid
 %description -n %{real_name}-pfcmd-suid
 The %{real_name}-pfcmd-suid is a C wrapper to replace perl-suidperl dependency.
 See https://bugzilla.redhat.com/show_bug.cgi?id=611009
+
+%package -n %{real_name}-config
+Group: System Environment/Daemons
+Requires: perl(Cache::BDB)
+Requires: perl(Log::Fast)
+AutoReqProv: 0
+Summary: Manage PacketFence Configuration
+BuildArch: noarch
+
+%description -n %{real_name}-config
+The %{real_name}-config is a daemon that manage PacketFence configuration.
 
 
 %prep
@@ -381,9 +393,11 @@ done
 %{__install} -d $RPM_BUILD_ROOT/usr/local/pf/var/rrd 
 %{__install} -d $RPM_BUILD_ROOT/usr/local/pf/var/session
 %{__install} -d $RPM_BUILD_ROOT/usr/local/pf/var/webadmin_cache
+%{__install} -d $RPM_BUILD_ROOT/usr/local/pf/var/control
 touch $RPM_BUILD_ROOT/usr/local/pf/var/cache_control
 cp Makefile $RPM_BUILD_ROOT/usr/local/pf/
 cp -r bin $RPM_BUILD_ROOT/usr/local/pf/
+cp -r addons/pfconfig/ $RPM_BUILD_ROOT/usr/local/pf/addons/
 cp -r addons/captive-portal/ $RPM_BUILD_ROOT/usr/local/pf/addons/
 cp -r addons/dev-helpers/ $RPM_BUILD_ROOT/usr/local/pf/addons/
 cp -r addons/high-availability/ $RPM_BUILD_ROOT/usr/local/pf/addons/
@@ -434,6 +448,9 @@ cp -r README $RPM_BUILD_ROOT/usr/local/pf/
 cp -r README.network-devices $RPM_BUILD_ROOT/usr/local/pf/
 cp -r UPGRADE.asciidoc $RPM_BUILD_ROOT/usr/local/pf/
 cp -r UPGRADE.old $RPM_BUILD_ROOT/usr/local/pf/
+#pfconfig
+%{__install} -D -m0755 addons/pfconfig/pfconfig.init $RPM_BUILD_ROOT%{_initrddir}/packetfence-config
+#end pfconfig
 # logfiles
 for LOG in %logfiles; do
     touch $RPM_BUILD_ROOT%logdir/$LOG
@@ -516,6 +533,14 @@ if ! /usr/bin/id pf &>/dev/null; then
         /usr/sbin/useradd -r -d "/usr/local/pf" -s /bin/sh -c "PacketFence" -M pf || \
                 echo Unexpected error adding user "pf" && exit
 fi
+
+%pre -n %{real_name}-config
+
+if ! /usr/bin/id pf &>/dev/null; then
+        /usr/sbin/useradd -r -d "/usr/local/pf" -s /bin/sh -c "PacketFence" -M pf || \
+                echo Unexpected error adding user "pf" && exit
+fi
+
 
 %post -n %{real_name}
 echo "Adding PacketFence startup script"
@@ -608,6 +633,10 @@ echo "Adding PacketFence remote Snort Sensor startup script"
 echo "Adding PacketFence remote ARP Sensor startup script"
 /sbin/chkconfig --add pfarp
 
+%post -n %{real_name}-config
+echo "Adding PacketFence config startup script"
+/sbin/chkconfig --add packetfence-config
+
 %preun -n %{real_name}
 if [ $1 -eq 0 ] ; then
         /sbin/service packetfence stop &>/dev/null || :
@@ -626,22 +655,41 @@ if [ $1 -eq 0 ] ; then
         /sbin/chkconfig --del pfarp
 fi
 
+%preun -n %{real_name}-config
+if [ $1 -eq 0 ] ; then
+        /sbin/service packetfence-config stop &>/dev/null || :
+        /sbin/chkconfig --del packetfence-config
+fi
+
 %postun -n %{real_name}
 if [ $1 -eq 0 ]; then
-        /usr/sbin/userdel pf || %logmsg "User \"pf\" could not be deleted."
-#       /usr/sbin/groupdel pf || %logmsg "Group \"pf\" could not be deleted."
-#else
-#       /sbin/service pf condrestart &>/dev/null || :
+        if /usr/bin/id pf &>/dev/null; then
+               /usr/sbin/userdel pf || %logmsg "User \"pf\" could not be deleted."
+#               /usr/sbin/groupdel pf || %logmsg "Group \"pf\" could not be deleted."
+#        else
+#               /sbin/service pf condrestart &>/dev/null || :
+        fi
 fi
 
 %postun -n %{real_name}-remote-snort-sensor
 if [ $1 -eq 0 ]; then
-        /usr/sbin/userdel pf || %logmsg "User \"pf\" could not be deleted."
+        if /usr/bin/id pf &>/dev/null; then
+                /usr/sbin/userdel pf || %logmsg "User \"pf\" could not be deleted."
+        fi
 fi
 
 %postun -n %{real_name}-remote-arp-sensor
 if [ $1 -eq 0 ]; then
-        /usr/sbin/userdel pf || %logmsg "User \"pf\" could not be deleted."
+        if /usr/bin/id pf &>/dev/null; then
+                /usr/sbin/userdel pf || %logmsg "User \"pf\" could not be deleted."
+        fi
+fi
+
+%postun -n %{real_name}-config
+if [ $1 -eq 0 ]; then
+        if /usr/bin/id pf &>/dev/null; then
+                /usr/sbin/userdel pf || %logmsg "User \"pf\" could not be deleted."
+        fi
 fi
 
 # TODO we should simplify this file manifest to the maximum keeping treating 
@@ -874,7 +922,41 @@ fi
 %dir                    /usr/local/pf/html/common
                         /usr/local/pf/html/common/*
                         /usr/local/pf/html/pfappserver/
+%config(noreplace)      /usr/local/pf/html/pfappserver/pfappserver.conf
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Admin.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/AdminRoles.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/Authentication.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/Authentication/Source.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/Fingerprints.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/Firewall_SSO.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/FloatingDevice.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/MacAddress.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/Networks.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/Pf.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/Profile/Default.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/Profile.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/Provisioning.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/Realm.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/Switch.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/System.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Configuration.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Configurator.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/UserAgents.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Config/Wrix.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/DB.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Graph.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Interface.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Node.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Roles.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Root.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/SavedSearch/Node.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/SavedSearch/User.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Service.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/SoH.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/User.pm
+%config(noreplace)      /usr/local/pf/html/pfappserver/lib/pfappserver/Controller/Violation.pm
                         /usr/local/pf/lib
+%exclude                /usr/local/pf/lib/pfconfig*
 %config(noreplace)      /usr/local/pf/lib/pf/billing/custom.pm
 %config(noreplace)      /usr/local/pf/lib/pf/floatingdevice/custom.pm
 %config(noreplace)      /usr/local/pf/lib/pf/inline/custom.pm
@@ -953,6 +1035,7 @@ fi
 %dir                    /usr/local/pf/var/rrd
 %dir                    /usr/local/pf/var/session
 %dir                    /usr/local/pf/var/webadmin_cache
+%dir                    /usr/local/pf/var/control
 %config(noreplace)      /usr/local/pf/var/cache_control
 
 # Remote snort sensor file list
@@ -982,7 +1065,22 @@ fi
 %files -n %{real_name}-pfcmd-suid
 %attr(6755, root, root) /usr/local/pf/bin/pfcmd
 
+%files -n %{real_name}-config
+%attr(0755, root, root) %{_initrddir}/packetfence-config
+%dir                    /usr/local/pf
+%dir                    /usr/local/pf/lib
+%dir                    /usr/local/pf/lib/pfconfig
+                        /usr/local/pf/lib/pfconfig/*
+%attr(0755, pf, pf)     /usr/local/pf/sbin/pfconfig
+%dir                    /usr/local/pf/addons/pfconfig
+%attr(0755, pf, pf)     /usr/local/pf/addons/pfconfig/cmd.pl
+%exclude                /usr/local/pf/addons/pfconfig/README.asciidoc
+%exclude                /usr/local/pf/addons/pfconfig/pfconfig.init
+
 %changelog
+* Fri Mar 06 2015 Inverse <info@inverse.ca> - 4.7.0-1
+- New release 4.7.0
+
 * Thu Feb 19 2015 Inverse <info@inverse.ca> - 4.6.1-1
 - New release 4.6.1
